clothing_shop/requirements/local.txt
asgiref==3.2.7
Django==3.0.6
django-model-utils==4.0.0
html5lib==1.0.1
Pillow==7.1.2
psycopg2==2.8.5
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.1
Unipath==1.1
webencodings==0.5.1
xhtml2pdf==0.2.4